pyproject.toml
[project]
name = "test-scrape-proj"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django==5.2.11",
    "django-celery-beat==2.8.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "httpx>=0.28.1",
    "redis>=7.2.0",
]
